Use int32 for the lower and upper for cross DEX compatiblity
Metropolis uses uint24, and Shadow uses int24.

To have a uniform interface (rebalance in IStrategy), accept int32 lower and upper bounds. Then they can safely be converted to int24 or uint24 without underflow or overflow issues

+++ contracts-metropolis/src/BaseVault.sol
@@ -225,10 +225,10 @@ abstract contract BaseVault is Clone, ERC20Upgradeable, ReentrancyGuardUpgradeab
      * @return low The lower bound of the range.
      * @return upper The upper bound of the range.
      */
-    function getRange() public view virtual override returns (uint24 low, uint24 upper) {
+    function getRange() public view virtual override returns (int32 low, int32 upper) {
         IStrategy strategy = _strategy;
 
-        return address(strategy) == address(0) ? (0, 0) : strategy.getRange();
+        return address(strategy) == address(0) ? (int32(0), int32(0)) : strategy.getRange();
     }
 
     /**
